Removed `critical` prop from `AbbreviatedCard` comopnent
