+++ client/header/activity-panel/panels/inbox/abbreviated-card.js
@@ -1,41 +1,19 @@
 /**
  * External dependencies
  */
-import { __, sprintf } from '@wordpress/i18n';
 import { Card, CardBody, Icon } from '@wordpress/components';
 import { Text } from '@woocommerce/experimental';
 import { Link } from '@woocommerce/components';
 import { recordEvent } from '@woocommerce/tracks';
 
-export const AbbreviatedCard = ( {
-	content,
-	critical,
-	icon,
-	href,
-	name,
-	title,
-} ) => {
+export const AbbreviatedCard = ( { content, icon, href, name, title } ) => {
 	const trackAbbreviatedCardClick = () => {
 		recordEvent( 'activity_panel_click', {
 			task: name,
 		} );
 	};
-	const addCriticalAlert = () => {
-		/* translators: Number of critical alerts */
-		const criticalAlertText =
-			critical > 1
-				? __( '%d critical alerts', 'woocommerce-admin' )
-				: __( '%d critical alert', 'woocommerce-admin' );
-		return (
-			<div>
-				<span
-					className={ 'woocommerce-abbreviated-card__critical-alert' }
-				>
-					{ sprintf( criticalAlertText, critical ) }
-				</span>
-			</div>
-		);
-	};
+
+	const isTextContent = typeof content === 'string';
 	return (
 		<Card className="woocommerce-abbreviated-card">
 			<CardBody size={ null }>
@@ -49,8 +27,7 @@ export const AbbreviatedCard = ( {
 					</div>
 					<div className="woocommerce-abbreviated-card__message">
 						<Text as="h3">{ title }</Text>
-						<Text>{ content }</Text>
-						{ critical > 0 && addCriticalAlert() }
+						{ isTextContent ? <Text>{ content }</Text> : content }
 					</div>
 				</Link>
 			</CardBody>
